+++ lib/auth/authclient/clt.go
@@ -1671,3 +1671,39 @@ type ClientI interface {
 	// GenerateAppToken creates a JWT token with application access.
 	GenerateAppToken(ctx context.Context, req types.GenerateAppTokenRequest) (string, error)
 }
+
+type CreateAppSessionForV15Client interface {
+	Ping(ctx context.Context) (proto.PingResponse, error)
+	CreateAppSession(ctx context.Context, req *proto.CreateAppSessionRequest) (types.WebSession, error)
+}
+
+// TryCreateAppSessionForClientCertV15 creates an app session if the auth
+// server is pre-v16 and returns the app session ID. This app session ID
+// is needed for user app certs requests before v16.
+// TODO (Joerger): DELETE IN v17.0.0
+func TryCreateAppSessionForClientCertV15(ctx context.Context, client CreateAppSessionForV15Client, username string, routeToApp proto.RouteToApp) (string, error) {
+	pingResp, err := client.Ping(ctx)
+	if err != nil {
+		return "", trace.Wrap(err)
+	}
+
+	// If the auth server is v16+, the client does not need to provide a pre-created app session.
+	const minServerVersion = "16.0.0-aa" // "-aa" matches all development versions
+	if utils.MeetsVersion(pingResp.ServerVersion, minServerVersion) {
+		return "", nil
+	}
+
+	ws, err := client.CreateAppSession(ctx, &proto.CreateAppSessionRequest{
+		Username:          username,
+		PublicAddr:        routeToApp.PublicAddr,
+		ClusterName:       routeToApp.ClusterName,
+		AWSRoleARN:        routeToApp.AWSRoleARN,
+		AzureIdentity:     routeToApp.AzureIdentity,
+		GCPServiceAccount: routeToApp.GCPServiceAccount,
+	})
+	if err != nil {
+		return "", trace.Wrap(err)
+	}
+
+	return ws.GetName(), nil
+}
